Typo fixes

+++ src/scripts/DefaultProperties.java
@@ -43,9 +43,12 @@ public class DefaultProperties {
 		// TODO: add main package properties
 		
 		/* network package properties */
-		props.setProperty("network::port", new Integer(1));
-		props.setProperty("network::timeout", new Integer(1));
-		props.setProperty("network::so_timeout", new Integer(1));
+		//Listening port
+		props.setProperty("network::port", new Integer(24600));
+		//Connection timeout, in seconds
+		props.setProperty("network::timeout", new Integer(4000));
+		//Network timeout, in milliseconds
+		props.setProperty("network::so_timeout", new Integer(35));
 		
 		/* operations package properties */
 		// TODO: add operations package properties
